requirements.txt
fastapi>=0.110.0,<0.112.0
uvicorn[standard]>=0.27.0,<0.29.0
PyJWT>=2.8.0,<3.0.0
passlib[bcrypt]>=1.7.4,<1.8.0
bcrypt>=4.1.2,<5.0.0
python-dotenv>=1.0.0,<2.0.0
asgiref>=3.7.2,<3.9.0
wfastcgi>=3.0.0,<4.0.0
python-multipart>=0.0.9,<0.0.10
